Move `CustomQueries` to `query` package
