+++ internal/sli/query/query_processing_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/keptn-contrib/dynatrace-service/internal/common"
 	"github.com/keptn-contrib/dynatrace-service/internal/credentials"
 	"github.com/keptn-contrib/dynatrace-service/internal/dynatrace"
-	"github.com/keptn-contrib/dynatrace-service/internal/keptn"
 	"github.com/keptn-contrib/dynatrace-service/internal/sli/result"
 	"github.com/keptn-contrib/dynatrace-service/internal/test"
 
@@ -199,10 +198,10 @@ func TestGetSLIValueWithOldAndNewCustomQueryFormat(t *testing.T) {
 	for _, testQuery := range testQueries {
 
 		customQueries := make(map[string]string)
-		customQueries[keptn.ResponseTimeP50] = testQuery
+		customQueries[ResponseTimeP50] = testQuery
 
-		p := createCustomQueryProcessing(t, keptnEvent, httpClient, keptn.NewCustomQueries(customQueries), timeframe)
-		sliResult := p.GetSLIResultFromIndicator(context.TODO(), keptn.ResponseTimeP50)
+		p := createCustomQueryProcessing(t, keptnEvent, httpClient, NewCustomQueries(customQueries), timeframe)
+		sliResult := p.GetSLIResultFromIndicator(context.TODO(), ResponseTimeP50)
 
 		assert.True(t, sliResult.Success())
 		assert.InDelta(t, 8.43340, sliResult.Value(), 0.001)
@@ -245,7 +244,7 @@ func runGetSLIResultFromIndicatorTest(t *testing.T, handler http.Handler) result
 
 	dh := createQueryProcessing(t, keptnEvent, httpClient, timeframe)
 
-	return dh.GetSLIResultFromIndicator(context.TODO(), keptn.ResponseTimeP50)
+	return dh.GetSLIResultFromIndicator(context.TODO(), ResponseTimeP50)
 }
 
 // Tests what happens when end time is too close to now. This test results in a short delay.
@@ -287,7 +286,7 @@ func TestGetSLISleep(t *testing.T) {
 
 	// time how long getting the SLI value takes
 	timeBeforeGetSLIValue := time.Now()
-	sliResult := dh.GetSLIResultFromIndicator(context.TODO(), keptn.ResponseTimeP50)
+	sliResult := dh.GetSLIResultFromIndicator(context.TODO(), ResponseTimeP50)
 	getSLIExectutionTime := time.Since(timeBeforeGetSLIValue)
 
 	assert.True(t, sliResult.Success())
@@ -309,7 +308,7 @@ func TestGetSLIValueWithErrorResponse(t *testing.T) {
 
 	dh := createQueryProcessing(t, keptnEvent, httpClient, timeframe)
 
-	sliResult := dh.GetSLIResultFromIndicator(context.TODO(), keptn.Throughput)
+	sliResult := dh.GetSLIResultFromIndicator(context.TODO(), Throughput)
 
 	assert.False(t, sliResult.Success())
 	assert.EqualValues(t, 0.0, sliResult.Value())
@@ -349,7 +348,7 @@ func TestGetSLIValueForIndicator(t *testing.T) {
 		customQueries := make(map[string]string)
 		customQueries[testConfig.indicator] = testConfig.query
 
-		ret := createCustomQueryProcessing(t, keptnEvent, httpClient, keptn.NewCustomQueries(customQueries), timeframe)
+		ret := createCustomQueryProcessing(t, keptnEvent, httpClient, NewCustomQueries(customQueries), timeframe)
 
 		sliResult := ret.GetSLIResultFromIndicator(context.TODO(), testConfig.indicator)
 
@@ -375,7 +374,7 @@ func TestGetSLIValueSupportsEnvPlaceholders(t *testing.T) {
 	customQueries := make(map[string]string)
 	customQueries[indicator] = "MV2;MicroSecond;entitySelector=type(SERVICE),tag(\"env_tag:$ENV.MY_ENV_TAG\")&metricSelector=builtin:service.response.time"
 
-	ret := createCustomQueryProcessing(t, keptnEvent, httpClient, keptn.NewCustomQueries(customQueries), timeframe)
+	ret := createCustomQueryProcessing(t, keptnEvent, httpClient, NewCustomQueries(customQueries), timeframe)
 	sliResult := ret.GetSLIResultFromIndicator(context.TODO(), indicator)
 
 	assert.True(t, sliResult.Success())
@@ -455,7 +454,7 @@ func TestGetSLIValueSupportsPlaceholders(t *testing.T) {
 		customQueries := make(map[string]string)
 		customQueries[testConfig.indicator] = testConfig.query
 
-		ret := createCustomQueryProcessing(t, keptnEvent, httpClient, keptn.NewCustomQueries(customQueries), timeframe)
+		ret := createCustomQueryProcessing(t, keptnEvent, httpClient, NewCustomQueries(customQueries), timeframe)
 
 		sliResult := ret.GetSLIResultFromIndicator(context.TODO(), testConfig.indicator)
 
@@ -469,11 +468,11 @@ func createQueryProcessing(t *testing.T, keptnEvent adapter.EventContentAdapter,
 		t,
 		keptnEvent,
 		httpClient,
-		keptn.NewEmptyCustomQueries(),
+		NewEmptyCustomQueries(),
 		timeframe)
 }
 
-func createCustomQueryProcessing(t *testing.T, keptnEvent adapter.EventContentAdapter, httpClient *http.Client, queries *keptn.CustomQueries, timeframe common.Timeframe) *Processing {
+func createCustomQueryProcessing(t *testing.T, keptnEvent adapter.EventContentAdapter, httpClient *http.Client, queries *CustomQueries, timeframe common.Timeframe) *Processing {
 	credentials, err := credentials.NewDynatraceCredentials("http://dynatrace", testDynatraceAPIToken)
 	assert.NoError(t, err)
